+++ src/js/components/index.js
@@ -1,14 +1,19 @@
 import accordion from '@uswds/uswds/js/usa-accordion';
 import banner from '@uswds/uswds/js/usa-banner';
+import button from '@uswds/uswds/js/usa-button';
 import characterCount from '@uswds/uswds/js/usa-character-count';
 import comboBox from '@uswds/uswds/js/usa-combo-box';
 import datePicker from '@uswds/uswds/js/usa-date-picker';
 import dateRangePicker from '@uswds/uswds/js/usa-date-range-picker';
 import fileInput from '@uswds/uswds/js/usa-file-input';
 import footer from '@uswds/uswds/js/usa-footer';
+import inPageNavigation from '@uswds/uswds/js/usa-in-page-navigation';
+import inputMask from '@uswds/uswds/js/usa-input-mask';
+import languageSelector from '@uswds/uswds/js/usa-language-selector';
 import modal from '@uswds/uswds/js/usa-modal';
 import header from '@uswds/uswds/js/usa-header';
 import password from '@uswds/uswds/js/_usa-password';
+import range from '@uswds/uswds/js/usa-range';
 import search from '@uswds/uswds/js/usa-search';
 import skipnav from '@uswds/uswds/js/usa-skipnav';
 import timePicker from '@uswds/uswds/js/usa-time-picker';
@@ -19,15 +24,20 @@ import validation from '@uswds/uswds/js/usa-validation';
 export {
   accordion,
   banner,
+  button,
   characterCount,
   comboBox,
   datePicker,
   dateRangePicker,
   fileInput,
   footer,
+  inPageNavigation,
+  inputMask,
+  languageSelector,
   modal,
   header,
   password,
+  range,
   search,
   skipnav,
   timePicker,