+++ typings/uswds__uswds.d.ts
@@ -1,0 +1,2 @@
+declare module '@uswds/uswds/js/usa-button';
+declare module '@uswds/uswds/js/usa-range';